+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Repository
 
 -   Update pre-commit hooks
+-   Update `thomaseizinger/create-pull-request` GiHub action
 
 ## [1.2.1] - 2021-07-19